refactor(perl-parser-core): split comprehensive_unit_tests.rs into per-domain test files
Split the 3916-line comprehensive_unit_tests.rs (36 nested mod blocks) into
36 individual top-level integration test files. Each file covers one domain:
recovery_parser, parser, parser_context, error_recovery, budget, token_stream,
trivia, position_mapper, line_index, ast_node, parse_output, source_location,
catastrophic_parse_error, error_context, error_classifier, missing_kind,
builtin_signatures, keyword_autoquoting, wave2a_qualified_subscripts,
bare_block_for_regression, and extended variants.

No behavior change — pure reorganization. All 1000+ tests pass.
New parser fix PRs now add a NEW file instead of touching a shared 4K-line file,
eliminating the primary source of merge conflicts in perl-parser-core.

Closes #1637
