+++ crates/perl-parser-core/tests/comprehensive_token_stream_extended_tests.rs
@@ -1,0 +1,58 @@
+use perl_parser_core::token_stream::TokenStream;
+
+#[test]
+fn on_stmt_boundary_resets_peek() -> Result<(), Box<dyn std::error::Error>> {
+    let mut stream = TokenStream::new("my $x; my $y;");
+    // Prime the peek
+    let _ = stream.peek();
+    // on_stmt_boundary should invalidate cached peeks
+    stream.on_stmt_boundary();
+    // After reset, peek should still work
+    if let Ok(token) = stream.peek() {
+        // We should get a valid token (the reparsed first token)
+        let _ = format!("{:?}", token.kind);
+    }
+    Ok(())
+}
+
+#[test]
+fn invalidate_peek_clears_cache() -> Result<(), Box<dyn std::error::Error>> {
+    let mut stream = TokenStream::new("1 + 2");
+    let _ = stream.peek();
+    let _ = stream.peek_second();
+    // Invalidate all cached peeks
+    stream.invalidate_peek();
+    // Should still work after invalidation
+    if let Ok(token) = stream.peek() {
+        let _ = format!("{:?}", token.kind);
+    }
+    Ok(())
+}
+
+#[test]
+fn peek_fresh_kind_on_empty() -> Result<(), Box<dyn std::error::Error>> {
+    let mut stream = TokenStream::new("");
+    let kind = stream.peek_fresh_kind();
+    // Should return Some(Eof) or similar
+    assert!(kind.is_some());
+    Ok(())
+}
+
+#[test]
+fn peek_fresh_kind_on_valid_input() -> Result<(), Box<dyn std::error::Error>> {
+    let mut stream = TokenStream::new("my $x;");
+    let kind = stream.peek_fresh_kind();
+    assert!(kind.is_some());
+    Ok(())
+}
+
+#[test]
+fn enter_format_mode_does_not_crash() -> Result<(), Box<dyn std::error::Error>> {
+    let mut stream = TokenStream::new("format STDOUT =\nsome text\n.\n");
+    stream.enter_format_mode();
+    // Should still be able to get tokens
+    if let Ok(token) = stream.peek() {
+        let _ = format!("{:?}", token.kind);
+    }
+    Ok(())
+}
